kharebot_Bot/requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
awscli==1.25.55
awsebcli==3.20.3
Babel==2.9.1
botocore==1.23.54
cachetools==4.2.2
cement==2.8.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
distlib==0.3.5
docutils==0.16
filelock==3.8.0
Flask==2.2.2
frozenlist==1.3.3
future==0.16.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
magic-filter==1.0.9
MarkupSafe==2.1.1
multidict==6.0.4
mysql-connector-python==8.0.30
pathspec==0.9.0
platformdirs==2.5.2
py-mon==1.1.0
pyasn1==0.4.8
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.11
virtualenv==20.16.3
wcwidth==0.1.9
Werkzeug==2.2.2
yarl==1.8.2
